fixes after review

+++ internal/core/main.go
@@ -73,6 +73,12 @@ func (o *Orchestrator) Stop() error {
 		return nil
 	}
 
+	if o.coreCmd.ProcessState.Exited() {
+		o.logger.Warn("process exited before stop was called")
+		o.coreCmd = nil
+		return nil
+	}
+
 	if err := o.coreCmd.Process.Signal(syscall.SIGTERM); err != nil {
 		return errors.Wrap(err, "failed to kill process")
 	}
@@ -143,7 +149,6 @@ func (o *Orchestrator) Run(ctx context.Context) error {
 
 			o.logger.WithField("task_id", taskID).Info("stopped process before executing task")
 			startDefaultMode, err := task.Execute(ctx)
-			fmt.Println("Task execution finished with startDefaultMode =", startDefaultMode)
 			if err != nil {
 				o.updateTaskFailed(taskID, err)
 				o.logger.WithError(err).
@@ -166,7 +171,6 @@ func (o *Orchestrator) Run(ctx context.Context) error {
 					return errors.Wrap(err, "failed to kill api process")
 				}
 			}
-			fmt.Println("UpdateStatus to completed")
 
 			// Update task status to Completed
 			if err = o.tasksDb.UpdateStatus(taskID, grpcTypes.ProcessStatus_PROCESS_STATUS_COMPLETED); err != nil {
